Allow to configure an http_adapter for Faraday.
This changeset extends the configuration of the library to add support
for an `http_adapter` option, which is passed down to Faraday and allows
to set a custom adapter to execute API calls.
It can be useful in circumstances where you want to switch using a
different, more performant or more customisable http client.